requirement.txt
Pillow
mysql-connector-python
babel
Babel
pyinstaller
